fix(swagger): fix security default value
And add a unit test on the generated swagger.json
